fix: check to see if the user exists to the express-validator, changed params to matchedData
